requirements.txt
beautifulsoup4
requests
pymysql
PySocks